+++ ToTe/Goal.h
@@ -13,7 +13,7 @@
 @property(nonatomic)int goal_id;
 @property(nonatomic)NSString *goal_title;
 @property(nonatomic)NSString *goal_description;
-@property(nonatomic)int goal_amount;
+@property(nonatomic)double goal_amount;
 @property(nonatomic)NSString *deadline;
 @property(nonatomic)NSString *goal_photo;
 @property(nonatomic)int priority;
@@ -21,13 +21,13 @@
 @property(nonatomic)double amount_tosave;
 @property(nonatomic)NSString* goal_start_date;
 
--(int)InsertGoal:(NSString *)g_title:(NSString *)g_description:(int)g_amount:(NSString *)deadline:(NSString *)g_photo:(double)amount_tosave;
+-(int)InsertGoal:(NSString *)g_title:(NSString *)g_description:(double)g_amount:(NSString *)deadline:(NSString *)g_photo:(double)amount_tosave;
 -(NSString *)ConvertDateFormat:(NSString *)end_date;
 -(NSString *)getCurrentDay;
 -(NSInteger)WeeksBetweenDate:(NSDate *)end_date;
 -(NSMutableArray *)SelectAllGoals;
 -(NSMutableArray *)SelectGoal:(int)g_id;
--(BOOL)UpdateGoal:(NSString *)g_title:(NSString *)g_description:(int)g_amount:(NSString *)deadline:(NSString *)g_photo:(double)amount_tosave:(int)g_id;
+-(BOOL)UpdateGoal:(NSString *)g_title:(NSString *)g_description:(double)g_amount:(NSString *)deadline:(NSString *)g_photo:(double)amount_tosave:(int)g_id;
 -(BOOL)DeleteGoal:(int)g_id;
 -(NSInteger)WeeksBetweenTwoDate:(NSDate *)start_date:(NSDate *)end_date;
 -(NSDate *)StringToDate:(NSString *)strDate;
